% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\docType{data}
\name{base_size}
\alias{base_size}
\title{Default base size for \code{\link{ggplot}} objects}
\format{
A numeric value
}
\usage{
base_size
}
\description{
Default base size for \code{\link{ggplot}} objects intended to render in lecture slides.
}
\keyword{datasets}
