% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{cfss_document}
\alias{cfss_document}
\title{cfss_document}
\usage{
cfss_document(inputFile, encoding)
}
\arguments{
\item{inputFile}{Rmd file to be rendered}

\item{encoding}{}
}
\description{
Function to generate slides/notes/etc. for class.
}

