% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ci}
\alias{ci}
\alias{add_ci}
\alias{plot_ci}
\title{Generate 95\% confidence intervals}
\usage{
add_ci(df_augment)

plot_ci(df_ci, x)
}
\arguments{
\item{df_augment}{Output of \code{\link[broom]{augment}}.}

\item{df_ci}{Output of \code{\link{add_ci}}.}

\item{x}{Name of variable to plot on the x-axis, passed as a string.}
}
\description{
Function to generate 95\% confidence intervals using the output
from \code{\link[broom]{augment}}. \code{\link{add_ci}} generates 95% confidence intervals
using the fitted values and standard errors, and \code{\link{plot_ci}} generates
a \code{\link[ggplot2]{ggplot}} line chart.
}
