% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cps}
\alias{cps}
\title{Chicago Public Schools school ratings}
\format{

}
\source{
\url{https://data.cityofchicago.org/Education/Chicago-Public-Schools-School-Progress-Reports-SY1/dw27-rash}
}
\usage{
cps
}
\description{
A dataset containing variables for Chicago Public Schools.
}
\keyword{datasets}
