% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dadmom}
\alias{dadmom}
\title{Dadmom}
\format{A data frame with 3 rows and 5 variables:
\describe{
  \item{famid}{Unit ID for family}
  \item{named}{Name of the dad}
  \item{incd}{Income of the dad}
  \item{namem}{Name of the mom}
  \item{incm}{Income of the mom}
}}
\source{
\url{http://www.ats.ucla.edu/stat/stata/modules/reshapel.htm}
}
\usage{
dadmom
}
\description{
A dataset for practicing converting from long-to-wide data form.
}
\keyword{datasets}
