% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{err-rate}
\alias{err-rate}
\alias{err.rate.rf}
\alias{err.rate.tree}
\title{Calculate error rate for a classification model}
\usage{
err.rate.rf(model, data)

err.rate.tree(model, data)
}
\arguments{
\item{model}{Model estimated using \code{\link[randomForest]{randomForest}}.}

\item{data}{Dataset used to calculate test error rate.}
}
\description{
\lifecycle{deprecated} \cr
Function to calculate error rate for a classification model.
}
