% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gun_deaths}
\alias{gun_deaths}
\title{Gun Deaths in America}
\format{A data frame with 100,798 rows and 10 variables:
\describe{
  \item{id}{Row number of the observation}
  \item{year}{Year the death occurred}
  \item{month}{Month the death occurred}
  \item{intent}{Intent of the death (homicide, suicide, accidental, unknown)}
  \item{police}{Indicator if the death was the result of an officer-involved shooting}
  \item{sex}{Victim's sex}
  \item{age}{Victim's age}
  \item{race}{Victim's race (White, Black, Hispanic, Asian/Pacific Islander,
  Native American/Native Alaskan)}
  \item{place}{Location where the shooting occurred}
  \item{education}{Highest level of education of the victim}
}}
\source{
\url{http://fivethirtyeight.com/features/gun-deaths/}
}
\usage{
gun_deaths
}
\description{
A dataset containing records on all firearm deaths in the United
States (excluding non-U.S. residents) from 2012-2014.
}
\keyword{datasets}
