% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{logit2prob}
\alias{logit2prob}
\title{Convert log-odds to predicted probabilities}
\usage{
logit2prob(x)
}
\arguments{
\item{x}{A vector of log-odds.}
}
\description{
Function to convert log-odds to predicted probabilities.
}
