% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelr.R
\name{model-quality-rcfss}
\alias{model-quality-rcfss}
\alias{mse}
\alias{mse.glm}
\title{Compute model quality for a given dataset}
\usage{
mse(model, data)

mse.glm(model, data)
}
\arguments{
\item{model}{A model}

\item{data}{The dataset}
}
\description{
\code{mse} is the mean-squared-error. Based on \code{\link[modelr]{rmse}}.
\code{mse.glm} works for \code{\link{glm}} models with binary response variables.
}
\examples{
mod <- lm(mpg ~ wt, data = mtcars)
mse(mod, mtcars)
}

