% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\alias{mse.data.frame}
\alias{mse_vec}
\title{Mean squared error}
\usage{
mse(data, ...)

\method{mse}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)

mse_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Mean squared error
}
