% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{prob2logodds}
\alias{prob2logodds}
\title{Convert probabilities to log-odds}
\usage{
prob2logodds(x)
}
\arguments{
\item{x}{A vector of predicted probabilities.}
}
\description{
Function to convert probabilities to log-odds.
}
