% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{prob2odds}
\alias{prob2odds}
\title{Convert predicted probabilities to log-odds}
\usage{
prob2odds(x)
}
\arguments{
\item{x}{A vector of predicted probabilities.}
}
\description{
\lifecycle{deprecated} \cr
Function to convert predicted probabilities to log-odds.
}
