% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{publishers}
\alias{publishers}
\title{Publishers}
\format{A data frame with 3 rows and 2 variables:
\describe{
  \item{publisher}{Name of publisher}
  \item{yr_founded}{Year the publisher was founded}
}}
\source{
\url{http://stat545.com/bit001_dplyr-cheatsheet.html}
}
\usage{
publishers
}
\description{
A dataset for practicing relational joins.
}
\keyword{datasets}

