% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scorecard}
\alias{scorecard}
\title{College affordability from College Scorecard}
\format{
A data frame with 1849 rows and 11 variables:
\describe{
  \item{unitid}{Unit ID for institution}
  \item{name}{Name of the college}
  \item{state}{State abbreviation}
  \item{type}{Type of college (Public; Private, nonprofit; Private, for-profit)}
  \item{admrate}{Undergraduate admissions rate (from 0-100\%)}
  \item{satavg}{Average SAT equivalent score of students admitted}
  \item{cost}{The average annual total cost of attendance, including tuition and fees,
  books and supplies, and living expenses}
  \item{netcost}{The average annual net cost of attendance (annual cost of attendance
  minus the average grant/scholarship aid)}
  \item{avgfacsal}{Average faculty salary (9 month)}
  \item{pctpell}{Percentage of undergraduates who receive a Pell Grant}
  \item{comprate}{Rate of first-time, full-time students at four-year institutions
  who complete their degree within six years}
  \item{firstgen}{Share of first-generation students}
  \item{debt}{Median debt of students after leaving school}
  \item{locale}{Locale of institution}
  \item{openadmp}{Open admissions policy indicator}
}
}
\source{
\url{https://collegescorecard.ed.gov/data/}
}
\usage{
scorecard
}
\description{
A dataset containing a sample of observations and variables
from the Department of Education's College Scorecard (data is for the 2018-19 academic year)
}
\keyword{datasets}
