% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{superheroes}
\alias{superheroes}
\title{Superheroes}
\format{A data frame with 7 rows and 4 variables:
\describe{
  \item{name}{Name of superhero}
  \item{alignment}{Is the superhero good or bad?}
  \item{gender}{Gender of superhero}
  \item{publisher}{Publisher which created the superhero}
}}
\source{
\url{http://stat545.com/bit001_dplyr-cheatsheet.html}
}
\usage{
superheroes
}
\description{
A dataset for practicing relational joins.
}
\keyword{datasets}

