% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USCongress}
\alias{USCongress}
\title{A sample dataset containing labeled bills from the United State Congress.}
\format{A \code{tibble} containing six columns.
\describe{
  \item{ID}{A unique identifier for the bill.}
  \item{cong}{The session of congress in which the bill first appeared.}
  \item{billnum}{The number of the bill as it appears in the congressional docket.}
  \item{h_or_sen}{A field specifying whether the bill was introduced in
  the House (HR) or the Senate (S).}
  \item{major}{A manually labeled topic code corresponding to the subject of the bill.}
  \item{text}{A short text description of the bill.}
}}
\source{
First included in the \href{https://github.com/timjurka/RTextTools}{\code{RTextTools}}
package (now deprecated). Original source is \url{http://www.congressionalbills.org/}.
}
\usage{
USCongress
}
\description{
A sample dataset containing labeled bills from the United States Congress,
compiled by Professor John D. Wilkerson at the University of Washington, Seattle
and E. Scott Adler at the University of Colorado, Boulder.
}
\keyword{datasets}
