% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worldbank}
\alias{worldbank}
\title{World Bank indicators}
\format{

}
\source{
\url{https://databank.worldbank.org/home.aspx}
}
\usage{
worldbank
}
\description{
A dataset containing both quantitative variables and qualitative variables for six countries from 2005 to 2017.
The countries included are: Argentina, China, United Kingdom, Indonesia, Norway, and United States
}
\keyword{datasets}
