% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{xaringan_wide}
\alias{xaringan_wide}
\title{xaringan_wide}
\usage{
xaringan_wide()
}
\description{
Function to generate slides for class using \code{\link[xaringan]{moon_reader}}.
Must pair with output from \code{xaringantheemer}
}
