% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worldbank}
\alias{worldbank}
\title{World Bank indicators}
\format{
A data frame with 78 rows and 14 variables:
\describe{
  \item{iso3c}{Country code ISO3.}
  \item{date}{Year in which the data is for.}
  \item{iso2c}{Country code ISO2.}
  \item{country}{Name of the country.}
  \item{perc_energy_fosfuel}{Fossil fuel energy consumption ( \% of total).}
  \item{rnd_gdpshare}{Research and development expenditure (\% of GDP).}
  \item{percgni_adj_gross_savings}{Adjusted savings: gross savings (\% of GNI).}
  \item{real_netinc_percap}{Adjusted net national income per capita (constant 2010 US$).}
  \item{gdp_capita}{GDP per capita (current USD).}
  \item{top10perc_incshare}{Income share held by highest 10 \%.}
  \item{employment_ratio}{Employment to population ratio, 15+, total (\%) (national estimate).}
  \item{life_exp}{Life expectancy at birth, total (years).}
  \item{pop_growth}{Population growth (annual \%).}
  \item{pop}{Total population.}
}
}
\source{
\url{https://databank.worldbank.org/home.aspx}
}
\usage{
worldbank
}
\description{
A dataset containing both quantitative variables and qualitative variables for six countries from 2005 to 2017.
The countries included are: Argentina, China, United Kingdom, Indonesia, Norway, and United States.
}
\keyword{datasets}
